requirements.txt
wxPython==4.1.1
requests==2.27.1
pyspellchecker==0.6.3
